[Chore] #1 - 폴더링
