+++ NotToDo/NotToDo/Network/Base/GeneralResponse.swift
@@ -1,0 +1,8 @@
+//
+//  GeneralResponse.swift
+//  NotToDo
+//
+//  Created by 강윤서 on 2022/12/27.
+//
+
+import Foundation
