+++ NotToDo/NotToDo/Presentation/HomeScene/ViewControllers/HomeViewController.swift
@@ -1,5 +1,5 @@
 //
-//  ViewController.swift
+//  HomeViewController.swift
 //  NotToDo
 //
 //  Created by 강윤서 on 2022/12/27.
@@ -7,13 +7,11 @@
 
 import UIKit
 
-class ViewController: UIViewController {
+class HomeViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        view.backgroundColor = .white
+
     }
 
-
 }
-
